api: move holdings-related functions to api.holdings
